Move relationship API to Route Handler
