+++ app/api/v2/utility.ts
@@ -1,0 +1,11 @@
+import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { User } from "@prisma/client";
+import { getServerSession } from "next-auth/next";
+
+export async function getUser(): Promise<User> {
+  const session = await getServerSession(authOptions);
+  if (!session || !session.user) {
+    throw Error("Unauthorized");
+  }
+  return session.user;
+}
