--- pages/api/v2/relationships/index.ts
@@ -1,41 +1,0 @@
-import { Relationship } from "@/datatypes/relationship";
-import { prisma } from "@/db";
-import { normalizeRelationship } from "@/selectors/relationships";
-import { User } from "@prisma/client";
-import { NextApiRequest, NextApiResponse } from "next";
-import { ApiErrorMethodUnknown } from "../types";
-import { withErrorHandling, withUser } from "../utility";
-
-export interface RelationshipsGetResult {
-  normalizedRelationships: Relationship[];
-}
-
-export default withErrorHandling(
-  withUser(async function handle(
-    req: NextApiRequest,
-    res: NextApiResponse,
-    user: User
-  ) {
-    if (req.method === "GET") {
-      const relationships = await prisma.relationship.findMany({
-        where: {
-          OR: [{ recipientId: user.id }, { senderId: user.id }],
-        },
-        include: {
-          recipient: true,
-          sender: true,
-        },
-      });
-
-      const normalizedRelationships = relationships.map((r) =>
-        normalizeRelationship(r, user.id)
-      );
-
-      res
-        .status(200)
-        .json({ normalizedRelationships } as RelationshipsGetResult);
-    } else {
-      throw new ApiErrorMethodUnknown();
-    }
-  })
-);
